small tooltip tweaks
