+++ app/components/shared/Tooltip/index.js
@@ -1,25 +1,30 @@
-import React from "react";
 import { Motion } from "react-motion";
-import state from "./state";
 import { injectState } from "freactal";
+import { Aux } from "../";
+import state from "./state";
 import "style/Tooltip.less";
 
 const px = n => n + "px";
-const mx = ({ top, left, opacity }) => ({ top: px(top), left: px(left), opacity });
+const mx = ({ top, left, opacity }, width) => ({ top: px(top), left: px(left), opacity, ...width });
 
-const Tooltip = ({ text, children, state: { position, startingPosition }, effects: { updatePosition, onMouseEnter, onMouseLeave }}) => {
+const Tooltip = ({ text, warning, disabled, className, tipWidth, children, state: { position }, effects: { updatePosition, onMouseEnter, onMouseLeave }}) => {
   let tooltip = null;
 
   const onMouseMove = ({ clientX, clientY }) => updatePosition(tooltip, clientX, clientY);
+  const container = ["tooltipContainer", className].join(" ");
+  const tip = ["tip", warning ? "warning" : null].join(" ");
+  const Wrapper = className ? "div" : Aux;
+  const width = tipWidth ? { width: tipWidth + "px" } : {};
 
-  return (
-    <div className="tooltipContainer" {...{ onMouseMove, onMouseEnter, onMouseLeave }} >
-      { children }
-      <Motion style={ position } defaultStyle={ startingPosition }>
-        { style => <div style={ mx(style) } className="tip" ref={ tip => tooltip = tip }>{ text }</div> }
+  return disabled ? <Wrapper className={ className }>{ children }</Wrapper> : (
+    <div className={ container } {...{ onMouseMove, onMouseEnter, onMouseLeave }} >
+        { children }
+      <Motion style={ position }>
+        { style => <div style={ mx(style, width) } className={ tip } ref={ tip => tooltip = tip }>{ text }</div> }
       </Motion>
     </div>
   );
 };
 
 export default state(injectState(Tooltip));
+
