+++ app/style/Tooltip.less
@@ -11,6 +11,7 @@
   box-shadow: 3px 3px 6px 0px rgba(0, 0, 0, 0.16);
   padding: 5px;
   z-index: 1000;
+  pointer-events: none;
 }
 
 .tip.warning {
